Height Measurement
This is an Arduino based project which is used to calculate the height
of any person accurately, both, in centimeters and inches. It uses
Ultrasonic sensor to calculate the height. This device may be fixed on
any wall at a certain height and then is calculates the reference height
from the ground. Once the refrence height is calculated, this device
displays the accurate height of the person standing beneath it. An LCD
is attached to controller to display the height.